Update Jogo_de_memoria.c
This update involves changing the logic. Previously, a sequence of six numbers was drawn and the user had to press the corresponding buttons to add them up with a kind of normalization. Now, this is no longer necessary, as the if structures select each button, so it's enough to add up the corresponding values of the pins. And so, we have two vectors with the same dimensions, being able to evaluate their values according to indices and do the appropriate treatments.