Update Jogo_de_memoria.c
Improving the code. The commit revolves around simplification. In other words, code that would normally be repeated has been made into a function to be called instead of this repetitive code. In addition to using registers to define the input and output of the pins and adding 0/1 to the Ports.