Update Memory_game.c
I've corrected some variables that were in the wrong positions, such as the states variable. In addition, I left the lines involving the lcd display with comments. This is because there wouldn't be too many wires in the project, and the lcd I made the code for depends on another type of library.